Documented PathFilter node.
